broke the huge AddDosage component into a smaller parent component with sub components
